Create Intel Core i9_arquitecture
In atomic quantum information devices, manipulation on qubit state can be enhanced by placing atoms into a cavity and hence coupled to coherent light. The dynamics is modeled as cavity quantum electrodynamics (CQED). Consider the qubit on a superconducting chip as an artificial ”atom”, AC signals propagating in the circuits as light. By coupling a superconducting qubit either capacitively or inductively to superconducting resonators, scientists have implemented the CQED method to manipulate quantum information on superconducting chips. Circuit QED is the formal terminology as the on-chip analogue of cavity QED.